alert week 3

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, Text, FlatList } from 'react-native';
+import { View, Text, FlatList, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux'
 import { baseUrl } from "../shared/baseUrl";
@@ -35,7 +35,22 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you want to delete the favorite dish ' + item.name + '?',
+                            [
+                                { text : 'Cancel', 
+                                    onPress: () => console.log(item.name + ' Not Deleted'),
+                                    style: 'cancel' 
+                                },
+                                { text : 'OK', 
+                                    onPress: () => this.props.deleteFavorite(item.id),
+                                }
+                            ],
+                            { cancelable : false } // you cant just dismiss the dialog; either press cancel or ok
+                        );
+                    } 
                 }
             ];
